backend/pyproject.toml
[project]
name = "flightradar"
version = "1.5.4"
description = "Flightradar is a Python backend for tracking and managing flight data."
requires-python = ">=3.11"
dependencies = [
    "apscheduler==3.11.0",
    "beautifulsoup4==4.14.2",
    "certifi==2025.10.5",
    "fastapi==0.118.2",
    "pydantic==2.12.0",
    "pymongo==4.15.3",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.2",
    "requests==2.32.5",
    "typer==0.19.2",
    "uvicorn==0.37.0",
    "python-dotenv==1.0.1",
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "protobuf>=4.25.0",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "slowapi==0.1.9",
    "fastapi-users[beanie]==14.0.1",
    "beanie>=1.25.0",
    "motor>=3.3.0"
]

[dependency-groups]
dev = [
    "requests-mock==1.8.0",
    "pytest==8.3.4",
    "pytest-cov==6.0.0"
]
